requirements.txt
google-cloud-bigquery
google-cloud-aiplatform
google-cloud-storage
pandas-gbq
shap
numpy==1.26.4
pandas==2.2.1
flask
joblib
scikit-learn==1.2.2
